# src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Maeven Collections API",
    "version": "1.0.0",
    "description": "API spec with endpoints grouped by tags."
  },
  "servers": [{ "url": "https://api.maevencollections.com" },
              { "url": "http://localhost:3000" }],
...
  "paths": {
...
    "/api/products/{productId}/variants": {
      "get": {
        "tags": ["Product Variants"],
        "summary": "List product variants by product ID",
        "parameters": [
          { "name": "productId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "is_active", "in": "query", "description": "Filter active/inactive variants", "schema": { "type": "boolean" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 100 } },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "default": 0 } }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
...
    "/api/orders/{orderId}/items": {
      "get": {
        "tags": ["Orders"],
        "summary": "List items for a specific order",
        "parameters": [
          { "name": "orderId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 100 } },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "default": 0 } }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean" },
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/OrderItem" } }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/gcash-fee-rules": {
      "get": { "tags": ["GCash Fee Rules"], "summary": "List GCash fee rules", "responses": { "200": { "description": "OK" } } },
      "post": {
        "tags": ["GCash Fee Rules"],
        "summary": "Create GCash fee rule",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/GCashFeeRuleCreate" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CreatedGCashFeeRuleResponse" }
              }
            }
          },
          "400": { "description": "Bad Request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },
          "500": { "description": "Server Error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } } }
        }
      }
    },
    "/api/gcash-fee-rules/{id}": {
      "get": { "tags": ["GCash Fee Rules"], "summary": "Get GCash fee rule", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "OK" } } },
      "put": { "tags": ["GCash Fee Rules"], "summary": "Update GCash fee rule", "responses": { "200": { "description": "Updated" } } },
      "delete": { "tags": ["GCash Fee Rules"], "summary": "Delete GCash fee rule", "responses": { "204": { "description": "No content" } } }
    },
    "/api/gcash-transactions": {
      "get": { "tags": ["GCash Transactions"], "summary": "List GCash transactions", "responses": { "200": { "description": "OK" } } },
      "post": {
        "tags": ["GCash Transactions"],
        "summary": "Create GCash transaction with fee lookup",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/GCashTransactionCreate" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CreatedGCashTransactionResponse" }
              }
            }
          },
          "400": { "description": "Bad Request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },
          "500": { "description": "Server Error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } } }
        }
      }
    },
    "/api/gcash-transactions/{id}": {
      "get": { "tags": ["GCash Transactions"], "summary": "Get GCash transaction", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "OK" } } }
    },
...
    "/api/prepaid-load-transactions": {
      "get": { "tags": ["Prepaid Load Transactions"], "summary": "List prepaid load transactions", "responses": { "200": { "description": "OK" } } },
      "post": {
        "tags": ["Prepaid Load Transactions"],
        "summary": "Create prepaid load transaction",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PrepaidLoadTransactionCreate" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CreatedPrepaidLoadTransactionResponse" }
              }
            }
          },
          "400": { "description": "Bad Request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },
          "500": { "description": "Server Error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } } }
        }
      }
    },
    "/api/prepaid-load-transactions/{id}": {
      "get": { "tags": ["Prepaid Load Transactions"], "summary": "Get prepaid load transaction", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "OK" } } }
    },
...
    "/api/purchase-orders-with-totals": {
      "get": {
        "tags": ["Purchase Orders"],
        "summary": "List purchase orders with item totals",
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 100 } },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "default": 0 } }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PurchaseOrdersWithTotalsResponse" }
              }
            }
          },
          "500": { "description": "Server Error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } } }
        }
      }
    },
    "/api/purchase-order-items": {
      "get": { "tags": ["Purchase Order Items"], "summary": "List purchase order items", "responses": { "200": { "description": "OK" } } },
      "post": {
        "tags": ["Purchase Order Items"],
        "summary": "Create purchase order item",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PurchaseOrderItemCreate" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CreatedPurchaseOrderItemResponse" }
              }
            }
          },
          "400": { "description": "Bad Request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },
          "500": { "description": "Server Error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } } }
        }
      }
    },
    "/api/purchase-order-items/{id}": {
      "get": { "tags": ["Purchase Order Items"], "summary": "Get purchase order item", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "OK" } } },
      "put": { "tags": ["Purchase Order Items"], "summary": "Update purchase order item", "responses": { "200": { "description": "Updated" } } },
      "delete": { "tags": ["Purchase Order Items"], "summary": "Delete purchase order item", "responses": { "204": { "description": "No content" } } }
    },
    "/api/purchase-orders/{poId}/estimates": {
      "get": {
        "tags": ["Purchase Order Estimates"],
        "summary": "List estimates for a purchase order",
        "parameters": [
          { "name": "poId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 100 } },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "default": 0 } }
        ],
        "responses": { "200": { "description": "OK" } }
      },
      "post": {
        "tags": ["Purchase Order Estimates"],
        "summary": "Create purchase order estimate",
        "parameters": [ { "name": "poId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } } ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PurchaseOrderEstimateCreate" } } }
        },
        "responses": { "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreatedPurchaseOrderEstimateResponse" } } } }, "400": { "description": "Bad Request" } }
      }
    },
    "/api/purchase-orders/{poId}/estimates/{id}": {
      "get": { "tags": ["Purchase Order Estimates"], "summary": "Get estimate", "parameters": [ { "name": "poId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }, { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } } ], "responses": { "200": { "description": "OK" } } },
      "put": { "tags": ["Purchase Order Estimates"], "summary": "Update estimate", "responses": { "200": { "description": "Updated" } } },
      "delete": { "tags": ["Purchase Order Estimates"], "summary": "Delete estimate", "responses": { "204": { "description": "No content" } } }
    },
    "/api/purchase-orders/{id}/variance": {
      "get": {
        "tags": ["Purchase Orders"],
        "summary": "Get estimated vs actual variance for a purchase order",
        "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } } ],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PurchaseOrderVariance" } } } } }
      }
    },
...
    "/api/reports/sales": {
      "get": {
        "tags": ["Reports"],
        "summary": "Sales summary grouped by period",
        "description": "Returns sales data (gross sales, net sales, discounts, items sold, order count) grouped by the chosen period. Defaults: daily → last 30 days, weekly → last 12 weeks, monthly → last 12 months, yearly → last 5 years.",
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "description": "Grouping period",
            "schema": { "type": "string", "enum": ["daily","weekly","monthly","yearly","custom"], "default": "daily" }
          },
          {
            "name": "from",
            "in": "query",
            "description": "Start date (YYYY-MM-DD). Required when period=custom.",
            "schema": { "type": "string", "format": "date", "example": "2026-06-01" }
          },
          {
            "name": "to",
            "in": "query",
            "description": "End date (YYYY-MM-DD, inclusive). Required when period=custom.",
            "schema": { "type": "string", "format": "date", "example": "2026-06-30" }
          },
          { "name": "branch_id", "in": "query", "description": "Filter by branch UUID", "schema": { "type": "string", "format": "uuid" } },
          { "name": "sales_channel_id", "in": "query", "description": "Filter by sales channel UUID", "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean", "example": true },
                    "period": { "type": "string", "example": "monthly" },
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/SalesPeriodItem" } }
                  }
                }
              }
            }
          },
          "400": { "description": "Bad Request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },
          "500": { "description": "Server Error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } } }
        }
      }
    },
    "/api/reports/sales/overview": {
      "get": {
        "tags": ["Reports"],
        "summary": "Sales KPI overview (single-row totals)",
        "description": "Returns aggregated totals for a date range — useful for dashboard KPI cards.",
        "parameters": [
          { "name": "from", "in": "query", "description": "Start date (YYYY-MM-DD)", "schema": { "type": "string", "format": "date", "example": "2026-07-01" } },
          { "name": "to", "in": "query", "description": "End date (YYYY-MM-DD, inclusive)", "schema": { "type": "string", "format": "date", "example": "2026-07-08" } },
          { "name": "branch_id", "in": "query", "schema": { "type": "string", "format": "uuid" } },
          { "name": "sales_channel_id", "in": "query", "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean", "example": true },
                    "data": { "$ref": "#/components/schemas/SalesOverview" }
                  }
                }
              }
            }
          },
          "400": { "description": "Bad Request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },
          "500": { "description": "Server Error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } } }
        }
      }
    },
    "/api/reports/sales/payments": {
      "get": {
        "tags": ["Reports"],
        "summary": "Payment method breakdown",
        "description": "Returns total collected amount and transaction count grouped by payment method for a given date range.",
        "parameters": [
          { "name": "from", "in": "query", "description": "Start date (YYYY-MM-DD)", "schema": { "type": "string", "format": "date", "example": "2026-07-01" } },
          { "name": "to", "in": "query", "description": "End date (YYYY-MM-DD, inclusive)", "schema": { "type": "string", "format": "date", "example": "2026-07-08" } },
          { "name": "branch_id", "in": "query", "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean", "example": true },
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/PaymentBreakdownItem" } }
                  }
                }
              }
            }
          },
          "400": { "description": "Bad Request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },
          "500": { "description": "Server Error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } } }
        }
      }
    },
    "/api/reports/sales/top-products": {
      "get": {
        "tags": ["Reports"],
        "summary": "Top-selling products by revenue",
        "description": "Returns the top N product variants ranked by total revenue within a date range.",
        "parameters": [
          { "name": "from", "in": "query", "description": "Start date (YYYY-MM-DD)", "schema": { "type": "string", "format": "date", "example": "2026-07-01" } },
          { "name": "to", "in": "query", "description": "End date (YYYY-MM-DD, inclusive)", "schema": { "type": "string", "format": "date", "example": "2026-07-08" } },
          { "name": "branch_id", "in": "query", "schema": { "type": "string", "format": "uuid" } },
          { "name": "limit", "in": "query", "description": "Max results (default 10, max 100)", "schema": { "type": "integer", "default": 10, "maximum": 100 } }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean", "example": true },
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/TopProductItem" } }
                  }
                }
              }
            }
          },
          "400": { "description": "Bad Request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },
          "500": { "description": "Server Error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } } }
        }
      }
    },
    "/api/reports/cash-reconciliation/daily": {
      "get": {
        "tags": ["Reports"],
        "summary": "Daily cash reconciliation report",
        "description": "Returns the cash reconciliation result for a branch and business date including opening cash, total sales, expected cash, actual cash, and shortage/overage.",
        "parameters": [
          { "name": "branch_id", "in": "query", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "business_date", "in": "query", "required": true, "schema": { "type": "string", "format": "date", "example": "2026-07-09" } }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean", "example": true },
                    "data": { "$ref": "#/components/schemas/CashReconciliationDailyReport" }
                  }
                }
              }
            }
          },
          "400": { "description": "Bad Request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },
          "404": { "description": "Not Found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },
          "500": { "description": "Server Error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } } }
        }
      }
    },
    "/api/cash-reconciliations": {
      "get": {
        "tags": ["Cash Reconciliation"],
        "summary": "List cash reconciliations",
        "parameters": [
          { "name": "branch_id", "in": "query", "schema": { "type": "string", "format": "uuid" } },
          { "name": "from", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "to", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 100 } },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "default": 0 } }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean", "example": true },
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/CashReconciliation" } }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/cash-reconciliations/{id}": {
      "get": {
        "tags": ["Cash Reconciliation"],
        "summary": "Get cash reconciliation by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean", "example": true },
                    "data": { "$ref": "#/components/schemas/CashReconciliationDailyReport" }
                  }
                }
              }
            }
          },
          "404": { "description": "Not Found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } } }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "properties": {
          "ok": { "type": "boolean", "example": false },
          "error": { "type": "string" }
        }
      },
...
      "OrderItem": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "order_id": { "type": "string", "format": "uuid" },
          "product_variant_id": { "type": "string", "format": "uuid" },
          "quantity": { "type": "integer" },
          "unit_price": { "type": "number", "format": "double" },
          "subtotal": { "type": "number", "format": "double" }
        }
      },
...
      "GCashFeeRuleCreate": {
        "type": "object",
        "properties": {
          "service_type": { "type": "string", "enum": ["cash_in","cash_out"] },
          "min_amount": { "type": "number", "format": "double" },
          "max_amount": { "type": "number", "format": "double", "nullable": true },
          "fee_amount": { "type": "number", "format": "double" },
          "is_active": { "type": "boolean", "default": true },
          "effective_from": { "type": "string", "format": "date-time", "nullable": true },
          "effective_to": { "type": "string", "format": "date-time", "nullable": true }
        },
        "required": ["service_type","min_amount","fee_amount"]
      },
...
      "CreatedGCashFeeRuleResponse": { "type": "object", "properties": { "ok": { "type": "boolean", "example": true }, "data": { "$ref": "#/components/schemas/GCashFeeRule" } } },
      "GCashTransactionCreate": {
        "type": "object",
        "properties": {
          "order_id": { "type": "string", "format": "uuid", "nullable": true },
          "branch_id": { "type": "string", "format": "uuid", "nullable": true },
          "customer_id": { "type": "string", "format": "uuid", "nullable": true },
          "service_type": { "type": "string", "enum": ["cash_in","cash_out"] },
          "principal_amount": { "type": "number", "format": "double" },
          "reference_number": { "type": "string" },
          "notes": { "type": "string", "nullable": true }
        },
        "required": ["service_type","principal_amount","reference_number"]
      },
...
      "CreatedGCashTransactionResponse": { "type": "object", "properties": { "ok": { "type": "boolean", "example": true }, "data": { "$ref": "#/components/schemas/GCashTransaction" } } },
...
      "CreatedPrepaidLoadTransactionResponse": {
        "type": "object",
        "properties": {
          "ok": { "type": "boolean", "example": true },
          "data": { "$ref": "#/components/schemas/PrepaidLoadTransaction" }
        }
      },
...
      "CreatedPurchaseOrderItemResponse": { "type": "object", "properties": { "ok": { "type": "boolean", "example": true }, "data": { "$ref": "#/components/schemas/PurchaseOrderItem" } } },
...
    }
  }
}